separate routes and database tests in different foldlders
